Refatoração para padrões novos

+++ product-api/src/main/java/br/com/udemy/productapi/modules/product/service/ProductService.java
@@ -1,5 +1,6 @@
 package br.com.udemy.productapi.modules.product.service;
 
+import static br.com.udemy.productapi.config.RequestUtil.getCurentRequest;
 import static org.springframework.util.ObjectUtils.isEmpty;
 
 import java.util.ArrayList;
@@ -9,6 +10,8 @@ import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import br.com.udemy.productapi.config.exception.SuccessResponse;
 import br.com.udemy.productapi.config.exception.ValidationException;
@@ -23,34 +26,73 @@ import br.com.udemy.productapi.modules.product.model.Product;
 import br.com.udemy.productapi.modules.product.repository.ProductRepository;
 import br.com.udemy.productapi.modules.sales.client.SalesClient;
 import br.com.udemy.productapi.modules.sales.dto.SalesConfirmationDTO;
+import br.com.udemy.productapi.modules.sales.dto.SalesProductResponse;
 import br.com.udemy.productapi.modules.sales.enums.SalesStatus;
 import br.com.udemy.productapi.modules.sales.rabbitmq.SalesConfirmationSender;
 import br.com.udemy.productapi.modules.supplier.service.SupplierService;
 import jakarta.transaction.Transactional;
+import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Service
+@AllArgsConstructor
 public class ProductService {
 
     private static final Integer ZERO = 0;
+    private static final String AUTHORIZATION = "Authorization";
+    private static final String TRANSACTION_ID = "transactionid";
+    private static final String SERVICE_ID = "serviceid";
 
-    @Autowired
-    private ProductRepository productRepository;
+    private final ProductRepository productRepository;
+    private final SupplierService supplierService;
+    private final CategoryService categoryService;
+    private final SalesConfirmationSender salesConfirmationSender;
+    private final SalesClient salesClient;
+    private final ObjectMapper objectMapper;
 
-    @Autowired
-    @Lazy
-    private CategoryService categoryService;
+    public ProductResponse save(ProductRequest request) {
+        validateProductDataInformed(request);
+        validateCategoryAndSupplierIdInformed(request);
+        var category = categoryService.findById(request.getCategoryId());
+        var supplier = supplierService.findById(request.getSupplierId());
+        var product = productRepository.save(Product.of(request, supplier, category));
+        return ProductResponse.of(product);
+    }
 
-    @Autowired
-    @Lazy
-    private SupplierService supplierService;
+    public ProductResponse update(ProductRequest request,
+            Integer id) {
+        validateProductDataInformed(request);
+        validateInformedId(id);
+        validateCategoryAndSupplierIdInformed(request);
+        var category = categoryService.findById(request.getCategoryId());
+        var supplier = supplierService.findById(request.getSupplierId());
+        var product = Product.of(request, supplier, category);
+        product.setId(id);
+        productRepository.save(product);
+        return ProductResponse.of(product);
+    }
 
-    @Autowired
-    private SalesConfirmationSender salesConfirmationSender;
+    private void validateProductDataInformed(ProductRequest request) {
+        if (isEmpty(request.getName())) {
+            throw new ValidationException("The product's name was not informed.");
+        }
+        if (isEmpty(request.getQuantityAvailable())) {
+            throw new ValidationException("The product's quantity was not informed.");
+        }
+        if (request.getQuantityAvailable() <= ZERO) {
+            throw new ValidationException("The quantity should not be less or equal to zero.");
+        }
+    }
 
-    @Autowired
-    private SalesClient salesClient;
+    private void validateCategoryAndSupplierIdInformed(ProductRequest request) {
+        if (isEmpty(request.getCategoryId())) {
+            throw new ValidationException("The category ID was not informed.");
+        }
+        if (isEmpty(request.getSupplierId())) {
+            throw new ValidationException("The supplier ID was not informed.");
+        }
+    }
 
     public List<ProductResponse> findAll() {
         return productRepository
@@ -62,7 +104,7 @@ public class ProductService {
 
     public List<ProductResponse> findByName(String name) {
         if (isEmpty(name)) {
-            throw new ValidationException("The supplier name must be informed");
+            throw new ValidationException("The product name must be informed.");
         }
         return productRepository
                 .findByNameIgnoreCaseContaining(name)
@@ -73,7 +115,7 @@ public class ProductService {
 
     public List<ProductResponse> findBySupplierId(Integer supplierId) {
         if (isEmpty(supplierId)) {
-            throw new ValidationException("The product's supplier Id must be informed");
+            throw new ValidationException("The product' supplier ID name must be informed.");
         }
         return productRepository
                 .findBySupplierId(supplierId)
@@ -84,7 +126,7 @@ public class ProductService {
 
     public List<ProductResponse> findByCategoryId(Integer categoryId) {
         if (isEmpty(categoryId)) {
-            throw new ValidationException("The product's category Id must be informed");
+            throw new ValidationException("The product' category ID name must be informed.");
         }
         return productRepository
                 .findByCategoryId(categoryId)
@@ -104,48 +146,6 @@ public class ProductService {
                 .orElseThrow(() -> new ValidationException("There's no product for the given ID."));
     }
 
-    public ProductResponse save(ProductRequest request) {
-        validateProductData(request);
-        validateCategoryAndSupplierInformed(request);
-        var category = categoryService.findById(request.getCategoryId());
-        var supplier = supplierService.findById(request.getSupplierId());
-        var product = productRepository.save(Product.of(request, supplier, category));
-        return ProductResponse.of(product);
-    }
-
-    public ProductResponse update(ProductRequest request, Integer id) {
-        validateProductData(request);
-        validateInformedId(id);
-        validateCategoryAndSupplierInformed(request);
-        var category = categoryService.findById(request.getCategoryId());
-        var supplier = supplierService.findById(request.getSupplierId());
-        var product = Product.of(request, supplier, category);
-        product.setId(id);
-        productRepository.save(product);
-        return ProductResponse.of(product);
-    }
-
-    private void validateProductData(ProductRequest request) {
-        if (isEmpty(request.getName())) {
-            throw new ValidationException("The product's name was not found");
-        }
-        if (isEmpty(request.getQuantityAvailable())) {
-            throw new ValidationException("The product's quantity was not informed");
-        }
-        if (isEmpty(request.getQuantityAvailable() <= ZERO)) {
-            throw new ValidationException("The quantity should not be less or equal to zero");
-        }
-    }
-
-    private void validateCategoryAndSupplierInformed(ProductRequest request) {
-        if (isEmpty(request.getCategoryId())) {
-            throw new ValidationException("The category id was not informed");
-        }
-        if (isEmpty(request.getSupplierId())) {
-            throw new ValidationException("The supplier id was not informed");
-        }
-    }
-
     public Boolean existsByCategoryId(Integer categoryId) {
         return productRepository.existsByCategoryId(categoryId);
     }
@@ -156,13 +156,20 @@ public class ProductService {
 
     public SuccessResponse delete(Integer id) {
         validateInformedId(id);
+        if (!productRepository.existsById(id)) {
+            throw new ValidationException("The product does not exists.");
+        }
+        var sales = getSalesByProductId(id);
+        if (!isEmpty(sales.getSalesIds())) {
+            throw new ValidationException("The product cannot be deleted. There are sales for it.");
+        }
         productRepository.deleteById(id);
-        return SuccessResponse.create("The product was deleted");
+        return SuccessResponse.create("The product was deleted.");
     }
 
     private void validateInformedId(Integer id) {
         if (isEmpty(id)) {
-            throw new ValidationException("The product ID must be informed");
+            throw new ValidationException("The supplier ID must be informed.");
         }
     }
 
@@ -170,46 +177,53 @@ public class ProductService {
         try {
             validateStockUpdateData(product);
             updateStock(product);
-        } catch (Exception e) {
-            log.error("Error while trying to update stock for message with error: {}", e.getMessage(), e);
-            var rejectedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.REJECTED);
+        } catch (Exception ex) {
+            log.error("Error while trying to update stock for message with error: {}", ex.getMessage(), ex);
+            var rejectedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.REJECTED,
+                    product.getTransactionid());
             salesConfirmationSender.sendSalesConfirmationMessage(rejectedMessage);
         }
     }
 
     private void validateStockUpdateData(ProductStockDTO product) {
-        if (isEmpty(product) || isEmpty(product.getSalesId())) {
-            throw new ValidationException("The product data or sales ID must be informed");
+        if (isEmpty(product)
+                || isEmpty(product.getSalesId())) {
+            throw new ValidationException("The product data and the sales ID must be informed.");
         }
-
         if (isEmpty(product.getProducts())) {
-            throw new ValidationException("The sales products must be informed");
+            throw new ValidationException("The sales' products must be informed.");
         }
-
-        product.getProducts().forEach(salesProduct -> {
-            if (isEmpty(salesProduct.getQuantity()) || isEmpty(salesProduct.getProductId())) {
-                throw new ValidationException("The productID and the quantity must be informed");
-            }
-        });
+        product
+                .getProducts()
+                .forEach(salesProduct -> {
+                    if (isEmpty(salesProduct.getQuantity())
+                            || isEmpty(salesProduct.getProductId())) {
+                        throw new ValidationException("The productID and the quantity must be informed.");
+                    }
+                });
     }
 
     @Transactional
     private void updateStock(ProductStockDTO product) {
         var productsForUpdate = new ArrayList<Product>();
-        product.getProducts().forEach(salesProduct -> {
-            var existingProduct = findById(salesProduct.getProductId());
-            validateQuantityInStock(salesProduct, existingProduct);
-            existingProduct.updateStock(salesProduct.getQuantity());
-            productsForUpdate.add(existingProduct);
-        });
+        product
+                .getProducts()
+                .forEach(salesProduct -> {
+                    var existingProduct = findById(salesProduct.getProductId());
+                    validateQuantityInStock(salesProduct, existingProduct);
+                    existingProduct.updateStock(salesProduct.getQuantity());
+                    productsForUpdate.add(existingProduct);
+                });
         if (!isEmpty(productsForUpdate)) {
             productRepository.saveAll(productsForUpdate);
-            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.APPROVED);
+            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.APPROVED,
+                    product.getTransactionid());
             salesConfirmationSender.sendSalesConfirmationMessage(approvedMessage);
         }
     }
 
-    private void validateQuantityInStock(ProductQuantityDTO salesProduct, Product existingProduct) {
+    private void validateQuantityInStock(ProductQuantityDTO salesProduct,
+            Product existingProduct) {
         if (salesProduct.getQuantity() > existingProduct.getQuantityAvailable()) {
             throw new ValidationException(
                     String.format("The product %s is out of stock.", existingProduct.getId()));
@@ -218,32 +232,59 @@ public class ProductService {
 
     public ProductSalesResponse findProductSales(Integer id) {
         var product = findById(id);
+        var sales = getSalesByProductId(product.getId());
+        return ProductSalesResponse.of(product, sales.getSalesIds());
+    }
+
+    private SalesProductResponse getSalesByProductId(Integer productId) {
         try {
-            var sales = salesClient.findSalesByProductId(product.getId())
-                    .orElseThrow(() -> new ValidationException("The sales was not found by this project."));
-            return ProductSalesResponse.of(product, sales.getSalesIds());
-        } catch (Exception e) {
-            throw new ValidationException("There was an error trying to get the product's sales.");
+            var currentRequest = getCurrentRequest();
+            var token = currentRequest.getHeader(AUTHORIZATION);
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Sending GET request to orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    productId, transactionid, serviceid);
+            var response = salesClient
+                    .findSalesByProductId(productId, token, transactionid)
+                    .orElseThrow(() -> new ValidationException("The sales was not found by this product."));
+            log.info("Recieving response from orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    objectMapper.writeValueAsString(response), transactionid, serviceid);
+            return response;
+        } catch (Exception ex) {
+            log.error("Error trying to call Sales-API: {}", ex.getMessage());
+            throw new ValidationException("The sales could not be found.");
         }
     }
 
     public SuccessResponse checkProductsStock(ProductCheckStockRequest request) {
-        if (isEmpty(request) || isEmpty(request.getProducts())) {
-            throw new ValidationException("The request data must be informed.");
+        try {
+            var currentRequest = getCurrentRequest();
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Request to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
+                    objectMapper.writeValueAsString(request), transactionid, serviceid);
+            if (isEmpty(request) || isEmpty(request.getProducts())) {
+                throw new ValidationException("The request data and products must be informed.");
+            }
+            request
+                    .getProducts()
+                    .forEach(this::validateStock);
+            var response = SuccessResponse.create("The stock is ok!");
+            log.info("Response to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
+                    objectMapper.writeValueAsString(response), transactionid, serviceid);
+            return response;
+        } catch (Exception ex) {
+            throw new ValidationException(ex.getMessage());
         }
-        request.getProducts()
-                .forEach(this::validateStock);
-        return SuccessResponse.create("The stock is OK.");
     }
 
     private void validateStock(ProductQuantityDTO productQuantity) {
         if (isEmpty(productQuantity.getProductId()) || isEmpty(productQuantity.getQuantity())) {
-            throw new ValidationException("Product ID and quantity must be informed");
+            throw new ValidationException("Product ID and quantity must be informed.");
         }
         var product = findById(productQuantity.getProductId());
         if (productQuantity.getQuantity() > product.getQuantityAvailable()) {
-            throw new ValidationException(String.format("The product %s is out of stock", product.getId()));
+            throw new ValidationException(String.format("The product %s is out of stock.", product.getId()));
         }
     }
-
 }
